requirements.txt
streamlit
yfinance
pandas
numpy
plotly
nltk
newsapi-python